requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
fastapi==0.114.0
frozenlist==1.4.1
future==1.0.0
h11==0.14.0
idna==3.8
line-bot-sdk==3.13.0
multidict==6.0.5
pydantic==2.9.1
pydantic_core==2.23.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
wrapt==1.16.0
yarl==1.11.0
openpyxl==3.1.2